+++ BiaoGe/Core/Module/MeetingHorn.lua
@@ -873,7 +873,7 @@ BG.Init2(function()
             BG.GameTooltip_Hide(edit)
         end
 
-        -- 发送
+        -- 右键菜单
         do
             local function GetWhisperText(onlylevel)
                 local text = " "
@@ -982,27 +982,16 @@ BG.Init2(function()
             -- 聊天框右键菜单
             if BG.IsNewUI then
                 if BiaoGe.options["MeetingHorn_whisper"] == 1 then
-                    if not BG.IsVanilla then
-                        Menu.ModifyMenu("MENU_UNIT_FRIEND", function(owner, rootDescription, contextData)
-                            rootDescription:CreateButton(L["装等+职业"], function()
-                                SendWhisper(contextData.name, "onlylevel")
-                            end)
-                        end)
-                    end
                     Menu.ModifyMenu("MENU_UNIT_FRIEND", function(owner, rootDescription, contextData)
                         rootDescription:CreateDivider()
                         -- rootDescription:CreateTitle("My Addon")
                         rootDescription:CreateButton(L["密语模板"], function()
                             SendWhisper(contextData.name)
                         end)
-                    end)
-                    if BG.IsVanilla then
-                        Menu.ModifyMenu("MENU_UNIT_FRIEND", function(owner, rootDescription, contextData)
-                            rootDescription:CreateButton(L["装等+职业"], function()
-                                SendWhisper(contextData.name, "onlylevel")
-                            end)
+                        rootDescription:CreateButton(L["装等+职业"], function()
+                            SendWhisper(contextData.name, "onlylevel")
                         end)
-                    end
+                    end)
                 end
             else
                 hooksecurefunc("UnitPopup_ShowMenu", function(arg1, which)