+++ BiaoGe/Core/Module/minimap.lua
@@ -17,17 +17,17 @@ local plugin = ldb:NewDataObject(AddonName, { text = AddonName, type = "data sou
 function plugin:OnClick(button) --function plugin.OnClick(self, button)
     if button == "LeftButton" then
         if IsControlKeyDown() then
-            BG.SetFBCD(nil, nil, true)
+            if SettingsPanel:IsVisible() then
+                HideUIPanel(SettingsPanel)
+            else
+                ns.InterfaceOptionsFrame_OpenToCategory("|cff00BFFFBiaoGe|r")
+                BG.MainFrame:Hide()
+            end
         else
             BG.MainFrame:SetShown(not BG.MainFrame:IsVisible())
         end
     elseif button == "RightButton" then
-        if SettingsPanel:IsVisible() then
-            HideUIPanel(SettingsPanel)
-        else
-            ns.InterfaceOptionsFrame_OpenToCategory("|cff00BFFFBiaoGe|r")
-            BG.MainFrame:Hide()
-        end
+        BG.SetFBCD(nil, nil, true)
     end
     BG.PlaySound(1)
 end