+++ GSE/CHANGELOG.md
@@ -1,6 +1,15 @@
 # GSE
 
-## [3.2.28-a](https://github.com/TimothyLuke/GSE-Advanced-Macro-Compiler/tree/3.2.28-a) (2025-04-25)
-[Full Changelog](https://github.com/TimothyLuke/GSE-Advanced-Macro-Compiler/compare/3.2.28...3.2.28-a) [Previous Releases](https://github.com/TimothyLuke/GSE-Advanced-Macro-Compiler/releases)
+## [3.2.29](https://github.com/TimothyLuke/GSE-Advanced-Macro-Compiler/tree/3.2.29) (2025-05-08)
+[Full Changelog](https://github.com/TimothyLuke/GSE-Advanced-Macro-Compiler/compare/3.2.28-a...3.2.29) [Previous Releases](https://github.com/TimothyLuke/GSE-Advanced-Macro-Compiler/releases)
 
-- #1790 Handle embedded CBOR encoding  
+- #1798 Loading into a Talent set sometimes doesnt update Actionbar Overrides with latest Sequence.  
+- #1797 Add rotation Assistant selected spell as last column in /gse debug  
+- #1796 Fix PTR Export to handle MoP API differences gracefully.  
+- #1795 #1794 LuaDoc update for C\_Pony changes  
+- #1795 #1794 LuaDoc update for C\_Pony changes  
+- #1795 MoP ToC updates  
+- #1794 Remove erroneas button assignment  
+- #1795 MoP initial API fixes.  
+- #1794 Initial 11.1.7 Updates  
+- #1791 Advanced Export only finds sequences for the current Class or Global even when all sequences are shown in the editor